backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.0
pydantic-settings>=2.0.3
sqlalchemy>=2.0.21
alembic>=1.12.0
psycopg2-binary>=2.9.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
elasticsearch==8.10.0
httpx>=0.25.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
email-validator>=2.0.0
websockets>=11.0.3